progress on XmlPerformanceParserTest. updated XmlFeatureParser and test data files to maintain 100% coverage
discovered the newline after a beginning tag counts as a 'characters' event; this was causing some tags like performanceSchedule to be processed by processChildEvent. moreover, actual unknown events, which were thought to be processed and ignored, were not being processed, a fact hidden by the processing of those known but irrelevant events. not a huge deal and probably would have no effect in practice, but I wanted to eliminate it.