Testing XmlSaver. added ChecksumVerifier and /checksum to test directory so that checksum can be used for regression testing.
ChecksumVerifier computes a checksum and compares it to a value in the file /checksum. if they don't match, it will spit out the computed checksum and the contents of the object. This is useful if the object is being tested for the first time or if it has changed and the checksum needs to be added to or updated in the file after a visual inspection of correctness. If the sums match, ChecksumVerifier is silent.